# `raw_detect` perf tweaks